Add DatastoreGarbageCollector class and make deleting more efficient
